requirements.txt
docxtpl
python-docx
google-generativeai
langgraph
langchain-core
python-dotenv
pytelegrambotapi
flask
